pyproject.toml
[project]
name = "ysearch-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "appnope==0.1.4",
    "asttokens==3.0.0",
    "babel==2.17.0",
    "beautifulsoup4==4.13.3",
    "blis==1.2.0",
    "catalogue==2.0.10",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "cloudpathlib==0.21.0",
    "comm==0.2.2",
    "confection==0.1.5",
    "contourpy==1.3.1",
    "courlan==1.3.2",
    "cycler==0.12.1",
    "cymem==2.0.11",
    "dateparser==1.2.1",
    "debugpy==1.8.12",
    "decorator==5.1.1",
    "duckdb==1.2.0",
    "elastic-transport==8.17.1",
    "elasticsearch==8.17.2",
    "executing==2.2.0",
    "faiss-cpu==1.10.0",
    "fastapi==0.115.8",
    "filelock==3.17.0",
    "fonttools==4.56.0",
    "fr-core-news-sm",
    "fsspec==2025.2.0",
    "h11==0.14.0",
    "htmldate==1.9.3",
    "huggingface-hub==0.29.1",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==8.32.0",
    "jedi==0.19.2",
    "jinja2==3.1.5",
    "joblib==1.4.2",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "justext==3.0.1",
    "kiwisolver==1.4.8",
    "langcodes==3.5.0",
    "langdetect==1.0.9",
    "language-data==1.3.0",
    "lxml==5.3.1",
    "lxml-html-clean==0.4.1",
    "marisa-trie==1.2.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.1",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "murmurhash==1.0.12",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "numpy==2.2.3",
    "opensearch-py",
    "packaging==24.2",
    "paho-mqtt>=2.1.0",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pika==1.3.2",
    "pillow==11.1.0",
    "platformdirs==4.3.6",
    "preshed==3.0.9",
    "prompt-toolkit==3.0.50",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pygments==2.19.1",
    "pymongo>=4.12.1",
    "pyparsing==3.2.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.1",
    "pyyaml==6.0.2",
    "pyzmq==26.2.1",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==13.9.4",
    "safetensors==0.5.2",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "sentence-transformers==3.4.1",
    "setuptools==75.8.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smart-open==7.1.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "spacy==3.8.4",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "srsly==2.5.1",
    "stack-data==0.6.3",
    "starlette==0.45.3",
    "sympy==1.13.1",
    "thinc==8.3.4",
    "threadpoolctl==3.5.0",
    "tld==0.13",
    "tokenizers==0.21.0",
    "torch==2.6.0",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "trafilatura==2.0.0",
    "traitlets==5.14.3",
    "transformers==4.49.0",
    "typer==0.15.2",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "tzlocal==5.3",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "warcio==1.7.5",
    "wasabi==1.1.3",
    "wcwidth==0.2.13",
    "weasel==0.4.1",
    "wrapt==1.17.2",
]

[tool.uv.sources]
fr-core-news-sm = { url = "https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.8.0/fr_core_news_sm-3.8.0-py3-none-any.whl" }
